feat: логика FilmController и UserController перемещена в сервисы
