+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -1,4 +1,5 @@
 package ru.yandex.practicum.filmorate.storage.film;
 
 public interface FilmStorage {
+
 }
