+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/InMemoryFilmStorage.java
@@ -3,5 +3,6 @@ package ru.yandex.practicum.filmorate.storage.film;
 import org.springframework.stereotype.Component;
 
 @Component
-public class InMemoryFilmStorage {
+public class InMemoryFilmStorage implements FilmStorage {
+
 }
