+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -3,5 +3,6 @@ package ru.yandex.practicum.filmorate.storage.user;
 import org.springframework.stereotype.Component;
 
 @Component
-public class InMemoryUserStorage {
+public class InMemoryUserStorage implements UserStorage {
+
 }
