+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -1,4 +1,5 @@
 package ru.yandex.practicum.filmorate.storage.user;
 
 public interface UserStorage {
+
 }
